Java application should run without internet connection
I've modified the code to get rid of spring boot. It's an 'usine à gaz'
for this application which is not needed.

* Cleaned up the RestEventStreamProvider class to use less dependencies
by using URL objects and jackson object mappers. I guess this was the
main reason to use spring before.

* Moved some code around and created DDD like packages for easier
understanding. The main entry point of the application is the Main
class, this one now instantiates the ProjectionRunner class which does
the logic.

* Modified the build script to load dependencies from a local directory.
This however implies that the dependencies are commited in the source control.
* Modified the gradle wrapper to load the gradle distribution from a
local file instead of the internet. This also means that the gradle
distribution is committed in the repository.

To use the offline mode simple run the following command: ./gradlew.bat
clean build --offline

If you add or remove dependencies you need to run the deleteLibs and copyToLibs
task to fetch the required dependencies. Don't forget to commit the
changes to the libs folder!